requirements.txt
opencv-python
torch
timm
geopandas
matplotlib
numpy
imgaug
